backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic-settings>=2.2.0
python-multipart>=0.0.9
httpx>=0.27.0
anthropic>=0.25.0
google-genai>=0.8.0
rembg>=2.0.57
Pillow>=10.3.0
numpy>=1.26.0
aiofiles>=23.2.1
python-dotenv>=1.0.1
onnxruntime>=1.19.0
